+++ client/src/components/UploadTasksButton.tsx
@@ -1,14 +1,12 @@
 import React, { useCallback, useState } from 'react'
 import { useUploadTasksMutation } from '../services/codefreak-api'
-import { Button, Icon, Modal, Spin, Upload } from 'antd'
+import { Alert, Button, Icon, Modal, Spin, Upload } from 'antd'
 import { RcFile } from 'antd/lib/upload/interface'
 
 const { Dragger } = Upload
 
 interface UploadTasksButtonProps {
-  onUploadCompleted: (
-    tasks: { id: string; createdAt: string; updatedAt: string }[]
-  ) => void
+  onUploadCompleted: (tasks: { id: string }[]) => void
 }
 
 const UploadTasksButton = (props: UploadTasksButtonProps) => {
@@ -28,9 +26,7 @@ const UploadTasksButton = (props: UploadTasksButtonProps) => {
           onUploadCompleted(
             r.data.uploadTasks.map(task => {
               return {
-                id: task.id,
-                createdAt: task.createdAt,
-                updatedAt: task.updatedAt
+                id: task.id
               }
             })
           )
@@ -41,8 +37,12 @@ const UploadTasksButton = (props: UploadTasksButtonProps) => {
   )
 
   const beforeUpload = useCallback(
-    (_: RcFile, fileList: RcFile[]) => {
-      onUpload(fileList)
+    (file: RcFile, fileList: RcFile[]) => {
+      // this function is called for every file, but the upload of the whole file list should run only once
+      if (fileList.indexOf(file) === 0) {
+        onUpload(fileList)
+      }
+
       return false
     },
     [onUpload]
@@ -63,6 +63,10 @@ const UploadTasksButton = (props: UploadTasksButtonProps) => {
           </Button>
         ]}
       >
+        <Alert
+          message="This action will create the imported tasks as new tasks and will not alter or delete existing tasks."
+          style={{ marginBottom: 16, marginTop: 16 }}
+        />
         <Dragger
           accept=".tar,.tar.gz,.zip"
           height={170}