+++ client/src/pages/task/TaskPoolPage.tsx
@@ -65,40 +65,13 @@ const TaskPoolPage: React.FC = () => {
 
   const update = () => result.refetch()
 
-  const handleUploadCompleted = (
-    createdTasks: { id: string; createdAt: string; updatedAt: string }[]
-  ) => {
+  const handleUploadCompleted = (createdTasks: { id: string }[]) => {
     if (createdTasks.length === 0) {
       messageService.error('Tasks could not be created')
     } else {
-      let numCreated = 0
-      let numUpdated = 0
-
-      createdTasks.forEach(createdTask => {
-        const existingTask = tasks.filter(task => task.id === createdTask.id)
-
-        if (existingTask.length > 0) {
-          const isTaskUpdated =
-            Date.parse(existingTask[0].updatedAt) <
-            Date.parse(createdTask.updatedAt)
-          if (isTaskUpdated) {
-            numUpdated++
-          }
-        } else {
-          numCreated++
-        }
-      })
-
-      if (numCreated + numUpdated > 0) {
-        if (numCreated > 0) {
-          messageService.success(`${numCreated} task(s) successfully created`)
-        }
-        if (numUpdated > 0) {
-          messageService.success(`${numUpdated} task(s) successfully updated`)
-        }
-      } else {
-        messageService.success('All tasks are up-to-date')
-      }
+      messageService.success(
+        `${createdTasks.length} task(s) successfully created`
+      )
     }
     result.refetch()
   }