+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * The task pool list and the 'add tasks to assignment' list are now sortable and filterable (#616)
 * The task pool can be exported and imported (#640)
 * Evaluation steps will now be canceled after a configurable timeout (#647)
+* Navigation can be hidden when giving assignment links to students (#667)
+* Add API for individual file operations (#666)
 
 ### Changed
 * Time limit can be specified on assignments and not on individual tasks (#635)